Initial attempt at making django_hstore work for django 1.7
