+++ django_hstore/managers.py
@@ -18,9 +18,11 @@ class HStoreManager(models.Manager):
     """
     use_for_related_fields = True
 
-    def get_query_set(self):
+    def get_queryset(self):
         return HStoreQuerySet(self.model, using=self._db)
-
+    
+    get_query_set = get_queryset
+    
     def hkeys(self, attr, **params):
         return self.filter(**params).hkeys(attr)
